researchagent-mcp/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.74.0
    # via researchagent-mcp (pyproject.toml)
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
arxiv==2.3.1
    # via researchagent-mcp (pyproject.toml)
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
cryptography==46.0.3
    # via pyjwt
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via anthropic
feedparser==6.0.12
    # via arxiv
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.12.0
    # via anthropic
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
mcp==1.21.2
    # via researchagent-mcp (pyproject.toml)
nest-asyncio==1.6.0
    # via researchagent-mcp (pyproject.toml)
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   anthropic
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pyjwt==2.10.1
    # via mcp
python-dotenv==1.2.1
    # via
    #   researchagent-mcp (pyproject.toml)
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via arxiv
rpds-py==0.29.0
    # via
    #   jsonschema
    #   referencing
sgmllib3k==1.0.0
    # via feedparser
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
sse-starlette==3.0.3
    # via mcp
starlette==0.50.0
    # via mcp
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   mcp
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via mcp
